Heroku config

+++ db/migrate/20160325062540_create_gasolinas.rb
@@ -2,11 +2,11 @@ class CreateGasolinas < ActiveRecord::Migration
   def change
     create_table :gasolinas do |t|
       t.string :clase
-      t.integer :galones_gravado
-      t.integer :precio_referencia
-      t.integer :porcentaje_alcohol
-      t.integer :base_gravable
-      t.integer :sobretasa
+      t.float :galones_gravado
+      t.float :precio_referencia
+      t.float :porcentaje_alcohol
+      t.float :base_gravable
+      t.float :sobretasa
       t.references :declaracion
       t.timestamps null: false
     end
